+++ Src/AwesomeAssertions/NumericAssertionsExtensions.cs
@@ -814,7 +814,7 @@ public static class NumericAssertionsExtensions
         if (assertion.Succeeded)
         {
             // ReSharper disable once PossibleInvalidOperationException
-            parent.BeApproximately(expectedValue.Value, precision, because, becauseArgs);
+            parent.BeApproximately(expectedValue!.Value, precision, because, becauseArgs);
         }
 
         return new AndConstraint<NullableNumericAssertions<float>>(parent);
@@ -952,7 +952,7 @@ public static class NumericAssertionsExtensions
         if (assertion.Succeeded)
         {
             // ReSharper disable once PossibleInvalidOperationException
-            parent.BeApproximately(expectedValue.Value, precision, because, becauseArgs);
+            parent.BeApproximately(expectedValue!.Value, precision, because, becauseArgs);
         }
 
         return new AndConstraint<NullableNumericAssertions<double>>(parent);
@@ -1092,7 +1092,7 @@ public static class NumericAssertionsExtensions
         if (assertion.Succeeded)
         {
             // ReSharper disable once PossibleInvalidOperationException
-            parent.BeApproximately(expectedValue.Value, precision, because, becauseArgs);
+            parent.BeApproximately(expectedValue!.Value, precision, because, becauseArgs);
         }
 
         return new AndConstraint<NullableNumericAssertions<decimal>>(parent);
@@ -1223,7 +1223,7 @@ public static class NumericAssertionsExtensions
         if (assertion.Succeeded)
         {
             // ReSharper disable once PossibleInvalidOperationException
-            parent.NotBeApproximately(unexpectedValue.Value, precision, because, becauseArgs);
+            parent.NotBeApproximately(unexpectedValue!.Value, precision, because, becauseArgs);
         }
 
         return new AndConstraint<NullableNumericAssertions<float>>(parent);
@@ -1355,7 +1355,7 @@ public static class NumericAssertionsExtensions
         if (assertionChain.Succeeded)
         {
             // ReSharper disable once PossibleInvalidOperationException
-            parent.NotBeApproximately(unexpectedValue.Value, precision, because, becauseArgs);
+            parent.NotBeApproximately(unexpectedValue!.Value, precision, because, becauseArgs);
         }
 
         return new AndConstraint<NullableNumericAssertions<double>>(parent);
@@ -1487,7 +1487,7 @@ public static class NumericAssertionsExtensions
         if (assertion.Succeeded)
         {
             // ReSharper disable once PossibleInvalidOperationException
-            parent.NotBeApproximately(unexpectedValue.Value, precision, because, becauseArgs);
+            parent.NotBeApproximately(unexpectedValue!.Value, precision, because, becauseArgs);
         }
 
         return new AndConstraint<NullableNumericAssertions<decimal>>(parent);