+++ Src/AwesomeAssertions/Primitives/StringAssertions.cs
@@ -182,16 +182,8 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
     public AndConstraint<TAssertions> NotBeEquivalentTo(string unexpected,
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            BeEquivalentTo(unexpected);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
         assertionChain
-            .ForCondition(notEquivalent)
+            .NotSatisfy(() => BeEquivalentTo(unexpected))
             .BecauseOf(because, becauseArgs)
             .FailWith("Expected {context:string} not to be equivalent to {0}{reason}, but they are.", unexpected);
 
@@ -220,16 +212,8 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
     {
         Guard.ThrowIfArgumentIsNull(config);
 
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            Subject.Should().BeEquivalentTo(unexpected, config);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
         assertionChain
-            .ForCondition(notEquivalent)
+            .NotSatisfy(() => Subject.Should().BeEquivalentTo(unexpected, config))
             .BecauseOf(because, becauseArgs)
             .FailWith("Expected {context:string} not to be equivalent to {0}{reason}, but they are.", unexpected);
 
@@ -1030,18 +1014,14 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
     {
         Guard.ThrowIfArgumentIsNull(unexpected, nameof(unexpected), "Cannot compare start of string with <null>.");
 
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            Subject.Should().StartWith(unexpected);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
         assertionChain
-            .ForCondition(Subject != null && notEquivalent)
             .BecauseOf(because, becauseArgs)
-            .FailWith("Expected {context:string} not to start with {0}{reason}, but found {1}.", unexpected, Subject);
+            .WithExpectation("Expected {context:string} not to start with {0}{reason}, ", unexpected, chain => chain
+                .ForCondition(Subject is not null)
+                .FailWith("but found <null>.")
+                .Then
+                .NotSatisfy(() => Subject.Should().StartWith(unexpected))
+                .FailWith("but found {0}.", unexpected, Subject));
 
         return new AndConstraint<TAssertions>((TAssertions)this);
     }
@@ -1126,19 +1106,14 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
     {
         Guard.ThrowIfArgumentIsNull(unexpected, nameof(unexpected), "Cannot compare start of string with <null>.");
 
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            Subject.Should().StartWithEquivalentOf(unexpected);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
         assertionChain
-            .ForCondition(Subject != null && notEquivalent)
             .BecauseOf(because, becauseArgs)
-            .FailWith("Expected {context:string} not to start with equivalent of {0}{reason}, but found {1}.", unexpected,
-                Subject);
+            .WithExpectation("Expected {context:string} not to start with equivalent of {0}{reason}, ", unexpected, chain => chain
+                .ForCondition(Subject is not null)
+                .FailWith("but found <null>.")
+                .Then
+                .NotSatisfy(() => Subject.Should().StartWithEquivalentOf(unexpected))
+                .FailWith("but found {0}.", Subject));
 
         return new AndConstraint<TAssertions>((TAssertions)this);
     }
@@ -1165,19 +1140,14 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
         Guard.ThrowIfArgumentIsNull(unexpected, nameof(unexpected), "Cannot compare start of string with <null>.");
         Guard.ThrowIfArgumentIsNull(config);
 
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            Subject.Should().StartWithEquivalentOf(unexpected, config);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
         assertionChain
-            .ForCondition(Subject != null && notEquivalent)
             .BecauseOf(because, becauseArgs)
-            .FailWith("Expected {context:string} not to start with equivalent of {0}{reason}, but found {1}.", unexpected,
-                Subject);
+            .WithExpectation("Expected {context:string} not to start with equivalent of {0}{reason}, but found <null>.", unexpected, chain => chain
+                .ForCondition(Subject is not null)
+                .FailWith(", but found <null>.")
+                .Then
+                .NotSatisfy(() => Subject.Should().StartWithEquivalentOf(unexpected, config))
+                .FailWith(", but found {0}.", unexpected, Subject));
 
         return new AndConstraint<TAssertions>((TAssertions)this);
     }
@@ -1227,18 +1197,14 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
     {
         Guard.ThrowIfArgumentIsNull(unexpected, nameof(unexpected), "Cannot compare end of string with <null>.");
 
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            Subject.Should().EndWith(unexpected);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
         assertionChain
-            .ForCondition(Subject != null && notEquivalent)
             .BecauseOf(because, becauseArgs)
-            .FailWith("Expected {context:string} not to end with {0}{reason}, but found {1}.", unexpected, Subject);
+            .WithExpectation("Expected {context:string} not to end with {0}{reason}, ", unexpected, chain => chain
+                .ForCondition(Subject is not null)
+                .FailWith("but found <null>.")
+                .Then
+                .NotSatisfy(() => Subject.Should().EndWith(unexpected))
+                .FailWith("but found {0}.", Subject));
 
         return new AndConstraint<TAssertions>((TAssertions)this);
     }
@@ -1323,18 +1289,14 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
     {
         Guard.ThrowIfArgumentIsNull(unexpected, nameof(unexpected), "Cannot compare end of string with <null>.");
 
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            Subject.Should().EndWithEquivalentOf(unexpected);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
         assertionChain
-            .ForCondition(Subject != null && notEquivalent)
             .BecauseOf(because, becauseArgs)
-            .FailWith("Expected {context:string} not to end with equivalent of {0}{reason}, but found {1}.", unexpected, Subject);
+            .WithExpectation("Expected {context:string} not to end with equivalent of {0}{reason}, ", unexpected, chain => chain
+                .ForCondition(Subject is not null)
+                .FailWith("but found <null>.")
+                .Then
+                .NotSatisfy(() => Subject.Should().EndWithEquivalentOf(unexpected))
+                .FailWith("but found {0}.", Subject));
 
         return new AndConstraint<TAssertions>((TAssertions)this);
     }
@@ -1361,18 +1323,14 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
         Guard.ThrowIfArgumentIsNull(unexpected, nameof(unexpected), "Cannot compare end of string with <null>.");
         Guard.ThrowIfArgumentIsNull(config);
 
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            Subject.Should().EndWithEquivalentOf(unexpected, config);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
         assertionChain
-            .ForCondition(Subject != null && notEquivalent)
             .BecauseOf(because, becauseArgs)
-            .FailWith("Expected {context:string} not to end with equivalent of {0}{reason}, but found {1}.", unexpected, Subject);
+            .WithExpectation("Expected {context:string} not to end with equivalent of {0}{reason}, ", unexpected, chain => chain
+                .ForCondition(Subject is not null)
+                .FailWith("but found <null>.")
+                .Then
+                .NotSatisfy(() => Subject.Should().EndWithEquivalentOf(unexpected, config))
+                .FailWith("but found {0}.", Subject));
 
         return new AndConstraint<TAssertions>((TAssertions)this);
     }
@@ -1782,24 +1740,13 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
         [StringSyntax("CompositeFormat")] string because = "", params object[] becauseArgs)
     {
         assertionChain
-            .ForCondition(!string.IsNullOrEmpty(unexpected) && Subject != null)
             .BecauseOf(because, becauseArgs)
-            .FailWith("Did not expect {context:string} to contain the equivalent of {0}{reason}, but found {1}.", unexpected,
-                Subject);
-
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            Subject.Should().ContainEquivalentOf(unexpected);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
-        assertionChain
-            .ForCondition(notEquivalent)
-            .BecauseOf(because, becauseArgs)
-            .FailWith("Did not expect {context:string} to contain the equivalent of {0}{reason} but found {1}.", unexpected,
-                Subject);
+            .WithExpectation("Did not expect {context:string} to contain the equivalent of {0}{reason}", unexpected, chain => chain
+                .ForCondition(!string.IsNullOrEmpty(unexpected) && Subject != null)
+                .FailWith(", but found {0}.", Subject)
+                .Then
+                .NotSatisfy(() => Subject.Should().ContainEquivalentOf(unexpected))
+                .FailWith(", but found {0}.", Subject));
 
         return new AndConstraint<TAssertions>((TAssertions)this);
     }
@@ -1825,24 +1772,13 @@ public class StringAssertions<TAssertions> : ReferenceTypeAssertions<string, TAs
         Guard.ThrowIfArgumentIsNull(config);
 
         assertionChain
-            .ForCondition(!string.IsNullOrEmpty(unexpected) && Subject != null)
             .BecauseOf(because, becauseArgs)
-            .FailWith("Did not expect {context:string} to contain the equivalent of {0}{reason}, but found {1}.", unexpected,
-                Subject);
-
-        bool notEquivalent;
-
-        using (var scope = new AssertionScope())
-        {
-            Subject.Should().ContainEquivalentOf(unexpected, config);
-            notEquivalent = scope.Discard().Length > 0;
-        }
-
-        assertionChain
-            .ForCondition(notEquivalent)
-            .BecauseOf(because, becauseArgs)
-            .FailWith("Did not expect {context:string} to contain the equivalent of {0}{reason}, but found {1}.", unexpected,
-                Subject);
+            .WithExpectation("Did not expect {context:string} to contain the equivalent of {0}{reason}", unexpected, chain => chain
+                .ForCondition(!string.IsNullOrEmpty(unexpected) && Subject != null)
+                .FailWith(", but found {0}.", Subject)
+                .Then
+                .NotSatisfy(() => Subject.Should().ContainEquivalentOf(unexpected, config))
+                .FailWith(", but found {0}.", Subject));
 
         return new AndConstraint<TAssertions>((TAssertions)this);
     }