+++ docs/_pages/releases.md
@@ -18,6 +18,7 @@ sidebar:
 ### Fixes
 * Use consistent newlines in failure messages - [#212](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/212)
 * Fix string equivalency check for custom comparer [#221](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/221)
+* Fix `EnumAssertions.BeOneOf` to use because description for `null` subject - [#268](https://github.com/AwesomeAssertions/AwesomeAssertions/pull/268)
 
 ## 9.1.0